# kat_rec_web/backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kat Rec Web Control Center API",
    "description": "Backend API for Kat Records automation workflow management",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/task/control": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Control Task",
        "description": "Control task for a channel or batch of channels\n\nActions:\n- start: Start a new task\n- pause: Pause current task\n- retry: Retry failed task\n- stop: Stop current task\n- retry_failed: Retry failed tasks for multiple channels (requires 'channels' field)\n\nExamples:\n- Single channel: {\"channel_id\": \"CH-001\", \"action\": \"start\"}\n- Batch retry: {\"action\": \"retry_failed\", \"channels\": [\"CH-006\", \"CH-009\"]}",
        "operationId": "control_task_api_task_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TaskControlRequest": {
        "properties": {
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "TaskControlRequest"
      },
      "TaskControlResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "timestamp"
        ],
        "title": "TaskControlResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}